Add files via upload
Admittedly it's been a very long time since I've looked at this code and while I know it's written in C++, requires compilation in CMake and uses VTK. I do not recall exactly what else is required to work for the environment. This did work and reached the end goal image also packaged in the ZIP file.